--- RetailCoder.VBE/Common/RawInput/MouseClickEventArgs.cs
@@ -1,14 +1,0 @@
-﻿using System;
-
-namespace RawInput_dll
-{
-    public class MouseClickEventArgs : EventArgs
-    {
-        public MouseClickEventArgs(MouseClickEvent arg)
-        {
-            MouseClickEvent = arg;
-        }
-        
-        public MouseClickEvent MouseClickEvent { get; private set; }
-    }
-}
