+++ RetailCoder.VBE/Common/WinAPI/BroadcastDeviceInterface.cs
@@ -1,0 +1,13 @@
+﻿using System;
+
+namespace Rubberduck.Common.WinAPI
+{
+    struct BroadcastDeviceInterface
+    {
+        public Int32 DbccSize;
+        public BroadcastDeviceType BroadcastDeviceType;
+        public Int32 DbccReserved;
+        public Guid DbccClassguid;
+        public char DbccName;
+    }
+}
