+++ RetailCoder.VBE/Common/WinAPI/BroadcastDeviceType.cs
@@ -1,0 +1,13 @@
+﻿namespace Rubberduck.Common.WinAPI
+{
+    enum BroadcastDeviceType
+    {
+        DBT_DEVTYP_OEM = 0,
+        DBT_DEVTYP_DEVNODE = 1,
+        DBT_DEVTYP_VOLUME = 2,
+        DBT_DEVTYP_PORT = 3,
+        DBT_DEVTYP_NET = 4,
+        DBT_DEVTYP_DEVICEINTERFACE = 5,
+        DBT_DEVTYP_HANDLE = 6,
+    }
+}
