+++ RetailCoder.VBE/Common/WinAPI/DeviceInfo.cs
@@ -1,0 +1,23 @@
+﻿using System.Runtime.InteropServices;
+
+namespace Rubberduck.Common.WinAPI
+{
+    [StructLayout(LayoutKind.Explicit)]
+    public struct DeviceInfo
+    {
+        [FieldOffset(0)]
+        public int Size;
+        [FieldOffset(4)]
+        public int Type;
+        [FieldOffset(8)]
+        public DeviceInfoMouse MouseInfo;
+        [FieldOffset(8)]
+        public DeviceInfoKeyboard KeyboardInfo;
+        [FieldOffset(8)]
+        public DeviceInfoHid HIDInfo;
+        public override string ToString()
+        {
+            return string.Format("DeviceInfo\n Size: {0}\n Type: {1}\n", Size, Type);
+        }
+    }
+}
