+++ RetailCoder.VBE/Common/WinAPI/DeviceInfoHid.cs
@@ -1,0 +1,22 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Rubberduck.Common.WinAPI
+{
+    public struct DeviceInfoHid
+    {
+        public uint VendorID;       // Vendor identifier for the HID
+        public uint ProductID;      // Product identifier for the HID
+        public uint VersionNumber;  // Version number for the device
+        public ushort UsagePage;    // Top-level collection Usage page for the device
+        public ushort Usage;        // Top-level collection Usage for the device
+
+        public override string ToString()
+        {
+            return string.Format("HidInfo\n VendorID: {0}\n ProductID: {1}\n VersionNumber: {2}\n UsagePage: {3}\n Usage: {4}\n", VendorID, ProductID, VersionNumber, UsagePage, Usage);
+        }
+    }
+}
