+++ RetailCoder.VBE/Common/WinAPI/DeviceType.cs
@@ -1,0 +1,9 @@
+﻿namespace Rubberduck.Common.WinAPI
+{
+    public enum DeviceType : uint
+    {
+        RIM_TYPE_MOUSE = 0,
+        RIM_TYPE_KEYBOARD = 1,
+        RIM_TYPE_HID = 2
+    }
+}
