+++ RetailCoder.VBE/Common/WinAPI/RawHID.cs
@@ -1,0 +1,17 @@
+﻿using System.Runtime.InteropServices;
+
+namespace Rubberduck.Common.WinAPI
+{
+    [StructLayout(LayoutKind.Sequential)]
+    internal struct RawHID
+    {
+        public uint dwSizHid;
+        public uint dwCount;
+        public byte bRawData;
+
+        public override string ToString()
+        {
+            return string.Format("Rawhib\n dwSizeHid : {0}\n dwCount : {1}\n bRawData : {2}\n", dwSizHid, dwCount, bRawData);
+        }
+    }
+}
