+++ RetailCoder.VBE/Common/WinAPI/RawInputDeviceInfo.cs
@@ -1,0 +1,9 @@
+﻿namespace Rubberduck.Common.WinAPI
+{
+    internal enum RawInputDeviceInfo : uint
+    {
+        RIDI_DEVICENAME = 0x20000007,
+        RIDI_DEVICEINFO = 0x2000000b,
+        PREPARSEDDATA = 0x20000005
+    }
+}
