+++ RetailCoder.VBE/Common/WinAPI/RawInputDeviceList.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Runtime.InteropServices;
+
+namespace Rubberduck.Common.WinAPI
+{
+    [StructLayout(LayoutKind.Sequential)]
+    internal struct RawInputDeviceList
+    {
+        public IntPtr hDevice;
+        public uint dwType;
+    }
+}
