+++ RetailCoder.VBE/Common/WinAPI/User32.cs
@@ -51,42 +51,6 @@ namespace Rubberduck.Common.WinAPI
         public delegate IntPtr WndProc(IntPtr hWnd, uint uMsg, IntPtr wParam, IntPtr lParam);
 
         /// <summary>
-        /// A pointer to the hook procedure. 
-        /// If the dwThreadId parameter is zero or specifies the identifier of a thread created by a different process, the lpfn parameter must point to a hook procedure in a DLL. 
-        /// Otherwise, lpfn can point to a hook procedure in the code associated with the current process.
-        /// </summary>
-        [DllImport("user32.dll", SetLastError = true)]
-        public static extern IntPtr SetWindowsHookEx(WindowsHook hookType, HookProc lpfn, IntPtr hMod, uint dwThreadId);
-        public delegate IntPtr HookProc(int code, IntPtr wParam, IntPtr lParam);
-
-        /// <summary>
-        /// Removes a hook procedure installed in a hook chain by the SetWindowsHookEx function.
-        /// </summary>
-        /// <param name="hhk">A handle to the hook to be removed. 
-        /// This parameter is a hook handle obtained by a previous call to SetWindowsHookEx.</param>
-        /// <returns>If the function succeeds, the return value is nonzero. 
-        /// If the function fails, the return value is zero. To get extended error information, call GetLastError.</returns>
-        [DllImport("user32.dll", SetLastError = true)]
-        [return: MarshalAs(UnmanagedType.Bool)]
-        public static extern bool UnhookWindowsHookEx(IntPtr hhk);
-
-        /// <summary>
-        /// Passes the hook information to the next hook procedure in the current hook chain. 
-        /// A hook procedure can call this function either before or after processing the hook information.
-        /// </summary>
-        /// <param name="hhk">This parameter is ignored.</param>
-        /// <param name="nCode">The hook code passed to the current hook procedure. 
-        /// The next hook procedure uses this code to determine how to process the hook information.</param>
-        /// <param name="wParam">The wParam value passed to the current hook procedure. 
-        /// The meaning of this parameter depends on the type of hook associated with the current hook chain.</param>
-        /// <param name="lParam">The lParam value passed to the current hook procedure. 
-        /// The meaning of this parameter depends on the type of hook associated with the current hook chain.</param>
-        /// <returns>This value is returned by the next hook procedure in the chain. 
-        /// The current hook procedure must also return this value. The meaning of the return value depends on the hook type.</returns>
-        [DllImport("user32.dll")]
-        public static extern IntPtr CallNextHookEx(IntPtr hhk, int nCode, IntPtr wParam, IntPtr lParam);
-
-        /// <summary>
         /// Retrieves a handle to the foreground window (the window with which the user is currently working). 
         /// The system assigns a slightly higher priority to the thread that creates the foreground window than it does to other threads.
         /// </summary>
@@ -191,6 +155,30 @@ namespace Rubberduck.Common.WinAPI
         [DllImport("user32.dll", ExactSpelling = true)]
         [return: MarshalAs(UnmanagedType.Bool)]
         public static extern bool KillTimer(IntPtr hWnd, IntPtr uIDEvent);
+
+        [DllImport("user32.dll", SetLastError = true)]
+        internal static extern int GetRawInputData(IntPtr hRawInput, DataCommand command, [Out] out InputData buffer, [In, Out] ref int size, int cbSizeHeader);
+
+        [DllImport("user32.dll", SetLastError = true)]
+        internal static extern int GetRawInputData(IntPtr hRawInput, DataCommand command, [Out] IntPtr pData, [In, Out] ref int size, int sizeHeader);
+
+        [DllImport("user32.dll", SetLastError = true)]
+        internal static extern uint GetRawInputDeviceInfo(IntPtr hDevice, RawInputDeviceInfo command, IntPtr pData, ref uint size);
+
+        [DllImport("user32.dll")]
+        private static extern uint GetRawInputDeviceInfo(IntPtr hDevice, uint command, ref DeviceInfo data, ref uint dataSize);
+
+        [DllImport("user32.dll", SetLastError = true)]
+        internal static extern uint GetRawInputDeviceList(IntPtr pRawInputDeviceList, ref uint numberDevices, uint size);
+
+        [DllImport("user32.dll", SetLastError = true)]
+        internal static extern bool RegisterRawInputDevices(RawInputDevice[] pRawInputDevice, uint numberDevices, uint size);
+
+        [DllImport("user32.dll", SetLastError = true)]
+        internal static extern IntPtr RegisterDeviceNotification(IntPtr hRecipient, IntPtr notificationFilter, DeviceNotification flags);
+
+        [DllImport("user32.dll", SetLastError = true)]
+        internal static extern bool UnregisterDeviceNotification(IntPtr handle);
 
         /// <summary>
         /// A helper function that returns <c>true</c> when the specified handle is that of the foreground window.
