+++ RetailCoder.VBE/Rubberduck.csproj
@@ -311,26 +311,46 @@
     <Compile Include="Common\Hotkeys\IHotkey.cs" />
     <Compile Include="Common\IRubberduckHooks.cs" />
     <Compile Include="Common\ModuleExporter.cs" />
-    <Compile Include="Common\RawInput\DataStructures.cs" />
-    <Compile Include="Common\RawInput\Enumerations.cs" />
-    <Compile Include="Common\RawInput\IRawDevice.cs" />
-    <Compile Include="Common\RawInput\KeyMapper.cs" />
-    <Compile Include="Common\RawInput\KeyPressEvent.cs" />
-    <Compile Include="Common\RawInput\KeyPressEventArgs.cs" />
-    <Compile Include="Common\RawInput\MouseClickEvent.cs" />
-    <Compile Include="Common\RawInput\MouseClickEventArgs.cs" />
-    <Compile Include="Common\RawInput\PreMessageFilter.cs" />
-    <Compile Include="Common\RawInput\RawInput.cs" />
-    <Compile Include="Common\RawInput\RawKeyboard.cs" />
-    <Compile Include="Common\RawInput\RawMouse.cs" />
-    <Compile Include="Common\RawInput\RegistryAccess.cs" />
-    <Compile Include="Common\RawInput\Win32.cs" />
+    <Compile Include="Common\WinAPI\EnumeratedDevice.cs" />
+    <Compile Include="Common\WinAPI\IRawDevice.cs" />
+    <Compile Include="Common\WinAPI\KeyMap.cs" />
+    <Compile Include="Common\WinAPI\RawDevice.cs" />
+    <Compile Include="Common\WinAPI\RawKeyEventArgs.cs" />
+    <Compile Include="Common\WinAPI\RawMouseEventArgs.cs" />
+    <Compile Include="Common\WinAPI\PreMessageFilter.cs" />
+    <Compile Include="Common\WinAPI\RawInput.cs" />
+    <Compile Include="Common\WinAPI\RawKeyboard.cs" />
+    <Compile Include="Common\WinAPI\RawMouse.cs" />
+    <Compile Include="Common\WinAPI\RegistryAccess.cs" />
+    <Compile Include="Common\WinAPI\Win32.cs" />
     <Compile Include="Common\RubberduckHooks.cs" />
     <Compile Include="Common\TimerHook.cs" />
+    <Compile Include="Common\WinAPI\BroadcastDeviceInterface.cs" />
+    <Compile Include="Common\WinAPI\BroadcastDeviceType.cs" />
+    <Compile Include="Common\WinAPI\DataCommand.cs" />
+    <Compile Include="Common\WinAPI\DeviceInfo.cs" />
+    <Compile Include="Common\WinAPI\DeviceInfoHid.cs" />
+    <Compile Include="Common\WinAPI\DeviceInfoKeyboard.cs" />
+    <Compile Include="Common\WinAPI\DeviceInfoMouse.cs" />
+    <Compile Include="Common\WinAPI\DeviceNotification.cs" />
+    <Compile Include="Common\WinAPI\DeviceType.cs" />
+    <Compile Include="Common\WinAPI\HidUsage.cs" />
+    <Compile Include="Common\WinAPI\HidUsagePage.cs" />
+    <Compile Include="Common\WinAPI\InputData.cs" />
     <Compile Include="Common\WinAPI\KBDLLHOOKSTRUCT.cs" />
     <Compile Include="Common\WinAPI\KBDLLHOOKSTRUCTFlags.cs" />
     <Compile Include="Common\WinAPI\MSLLHOOKSTRUCT.cs" />
     <Compile Include="Common\WinAPI\POINT.cs" />
+    <Compile Include="Common\WinAPI\RawData.cs" />
+    <Compile Include="Common\WinAPI\RawHID.cs" />
+    <Compile Include="Common\WinAPI\RawInputDevice.cs" />
+    <Compile Include="Common\WinAPI\RawInputDeviceFlags.cs" />
+    <Compile Include="Common\WinAPI\RawInputDeviceInfo.cs" />
+    <Compile Include="Common\WinAPI\RawInputDeviceList.cs" />
+    <Compile Include="Common\WinAPI\RawInputHeader.cs" />
+    <Compile Include="Common\WinAPI\RawKeyboardSTRUCT.cs" />
+    <Compile Include="Common\WinAPI\RawMouseSTRUCT.cs" />
+    <Compile Include="Common\WinAPI\UsButtonFlags.cs" />
     <Compile Include="Common\WinAPI\WindowsHook.cs" />
     <Compile Include="Common\WinAPI\Kernel32.cs" />
     <Compile Include="Common\KeyHookEventArgs.cs" />
